add designate contract
