+++ src/neo/SmartContract/Native/NativeContract.cs
@@ -25,6 +25,7 @@ namespace Neo.SmartContract.Native
         public static GasToken GAS { get; } = new GasToken();
         public static PolicyContract Policy { get; } = new PolicyContract();
         public static OracleContract Oracle { get; } = new OracleContract();
+        public static DesignateContract Designate { get; } = new DesignateContract();
 
         [ContractMethod(0, CallFlags.None)]
         public abstract string Name { get; }
